Move piece movement tests into its own file
